[Http] Move Routing Data class to Data namespace.
