+++ src/Valkyrja/Application/Valkyrja.php
@@ -23,7 +23,7 @@ use Valkyrja\Cli\Routing\Data\Data as CliData;
 use Valkyrja\Container\Data\Data as ContainerData;
 use Valkyrja\Container\Manager\Contract\Container;
 use Valkyrja\Event\Data\Data as EventData;
-use Valkyrja\Http\Routing\Data as HttpData;
+use Valkyrja\Http\Routing\Data\Data as HttpData;
 
 /**
  * Class Valkyrja.
