+++ src/Valkyrja/Http/Routing/Collection/Contract/Collection.php
@@ -14,8 +14,8 @@ declare(strict_types=1);
 namespace Valkyrja\Http\Routing\Collection\Contract;
 
 use Valkyrja\Http\Message\Enum\RequestMethod;
-use Valkyrja\Http\Routing\Data;
 use Valkyrja\Http\Routing\Data\Contract\Route;
+use Valkyrja\Http\Routing\Data\Data;
 
 /**
  * Interface Collection.
